backend_backup/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
async-timeout==4.0.3
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.60
langchain-text-splitters==0.3.8
langsmith==0.3.42
orjson==3.10.18
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
python-dotenv==1.1.0
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==9.1.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
zstandard==0.23.0
passlib[bcrypt]
python-jose[cryptography]
